move build_glob_set and SupportedFileType into file_utils
